requirements.txt
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==20.0
pathlib==1.0.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.2
python-slugify==4.0.0
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
text-unidecode==1.3
urllib3==1.25.7
wcwidth==0.1.8
zipp==0.6.0
